upload the RHUI ISO to the HAProxy node too
It's necessary to install puppet on the HAProxy node. Luckily, puppet is
in the rh-common repo in AWS, so everything has worked for us so far,
but if rhui3-automation is used outside AWS, the HAProxy setup fails.
Also, it's generally not nice to rely on unrelated repos that happen to
contain the packages that RHUI needs.