complete fully automated beta testing using Ansible
Set a long SSH connection timeout in the configuration. Without it, if
using a tunnel to an internal repo, the connections get closed in 60
seconds and the internal repo becomes unusable, breaking package
installations performed by tests.
Also disable host key checking so that Ansible can connect to new hosts
non-interactively.
Move the Ansible configuration file to the top-level directory so its
location does not have to be specified as an environment variable.
Run a simple command on RHUI nodes in order to establish SSH connections
to them if running tests is the only task. Necessary if using a tunnel
to an internal repo.